+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Show `Stdlib.Null` and `Stdlib.Nullable` completions for `Stdlib.null<'a>` and `Stdlib.nullable<'a>` types, respectively. https://github.com/rescript-lang/rescript/pull/7826
 - Fix generation of interfaces for module types containing multiple type constraints. https://github.com/rescript-lang/rescript/pull/7825
 - JSX preserve mode: fix "make is not a valid component name". https://github.com/rescript-lang/rescript/pull/7831
+- Rewatch: include parser arguments of experimental features. https://github.com/rescript-lang/rescript/pull/7836
 
 #### :memo: Documentation